+++ tests/location/test_location_location.py
@@ -34,6 +34,7 @@ def test_location_retrieve_all():
 
     assert locations, "No locations could be retrieved"
     assert "id" in locations[0], "'id' was not returned for the first list item"
+    assert "venue" in locations[0], "'venue' was not returned for the first list item"
 
 
 def test_location_retrieve_all_details():
@@ -43,6 +44,7 @@ def test_location_retrieve_all_details():
 
     assert locations, "No locations could be retrieved"
     assert "id" in locations[0], "'id' was not returned for first list item"
+    assert "venue" in locations[0], "'venue' was not returned for the first list item"
     assert (
         "recordings" in locations[0]
     ), "'recordings' was not returned for the first list item"
@@ -76,6 +78,8 @@ def test_location_retrieve_by_id(location_id: int):
 
     assert info, f"Location ID {location_id} not found"
     assert "venue" in info, f"'venue' was not returned for ID {location_id}"
+    assert "latitude" in info, f"'latitude' was not returned for ID {location_id}"
+    assert "longitude" in info, f"'longitude' was not returned for ID {location_id}"
 
 
 @pytest.mark.parametrize("location_id", [95])
@@ -89,6 +93,8 @@ def test_location_retrieve_details_by_id(location_id: int):
 
     assert info, f"Location ID {location_id} not found"
     assert "venue" in info, f"'venue' was not returned for ID {location_id}"
+    assert "latitude" in info, f"'latitude' was not returned for ID {location_id}"
+    assert "longitude" in info, f"'longitude' was not returned for ID {location_id}"
     assert "recordings" in info, f"'recordings' was not returned for ID {location_id}"
 
 
@@ -104,6 +110,8 @@ def test_location_retrieve_by_slug(location_slug: str):
 
     assert info, f"Location slug {location_slug} not found"
     assert "venue" in info, f"'venue' was not returned for slug {location_slug}"
+    assert "latitude" in info, f"'latitude' was not returned for ID {location_slug}"
+    assert "longitude" in info, f"'longitude' was not returned for ID {location_slug}"
 
 
 @pytest.mark.parametrize("location_slug", ["the-chicago-theatre-chicago-il"])
@@ -118,6 +126,8 @@ def test_location_retrieve_details_by_slug(location_slug: str):
 
     assert info, f"Location slug {location_slug} not found"
     assert "venue" in info, f"'venue' was not returned for slug {location_slug}"
+    assert "latitude" in info, f"'latitude' was not returned for ID {location_slug}"
+    assert "longitude" in info, f"'longitude' was not returned for ID {location_slug}"
     assert (
         "recordings" in info
     ), f"'recordings' was not returned for slug {location_slug}"